# openapi.yaml
openapi: 3.1.0
info:
  title: Synvera API
  version: 3.1.0
  description: >
    Composition-builder API for the Synvera medical procedure valuation platform.
    Implements CBHPM 2022 items 4.1, 4.2, and 5.1 for correct multi-procedure
    discounting and auxiliary surgeon remuneration.
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
...
  /api/calculate:
    post:
      summary: Calculate total valuation for a composition
      description: >
        Accepts a physician-assembled composition of CBHPM codes plus staffing
        configuration and access route type, then returns a detailed breakdown
        implementing CBHPM 2022 items 4.1/4.2 (multi-procedure discounting) and
        item 5.1 (auxiliary surgeon remuneration at 60/40/30/30%).
      operationId: calculate
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CalculateRequest'
      responses:
        '200':
          description: Detailed valuation breakdown.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CalculateResponse'
        '400':
          description: Invalid request body.
...
components:
  schemas:
...
    BillingMode:
      type: string
      enum: [PER_PROCEDURE, PER_SEGMENT, PER_VERTEBRA, PER_STRUCTURE]
      description: >
        How a procedure's base value scales with quantity.
        - PER_PROCEDURE: No quantity multiplier (default).
        - PER_SEGMENT: Value is multiplied by number of segments treated.
        - PER_VERTEBRA: Value is multiplied by number of vertebrae treated.
        - PER_STRUCTURE: Value is multiplied by number of structures treated.
    Specialty:
      type: string
      enum: [NEUROSURGERY, SPINE]
      description: >
        Medical domain classification.
        - NEUROSURGERY: General neurosurgery (default).
        - SPINE: Spine-specific procedures with billing variables.
    Laterality:
      type: string
      enum: [UNILATERAL, BILATERAL]
      description: >
        Whether procedure is billed for one side or both.
        - UNILATERAL: One side (multiplier = 1).
        - BILATERAL: Both sides (multiplier = 2, if applicable).
...
    AccessRouteType:
      type: string
      enum: [same, different]
      description: >
        Whether all selected procedures were performed via the same access route
        (CBHPM 4.1 – 50% discount on secondary procedures) or different access
        routes (CBHPM 4.2 – 70% discount on secondary procedures).
    SelectedCode:
      type: object
      required: [cbhpm_code, description, porte, billing_mode, specialty, laterality_support]
      properties:
        cbhpm_code:
          type: string
          description: CBHPM code selected by the physician.
        description:
          type: string
          description: CBHPM description (passed through from the procedure detail).
        porte:
          type: string
          description: Porte for this code (from the SBN catalog, not editable by the physician).
        billing_mode:
          $ref: '#/components/schemas/BillingMode'
          description: Billing mode inherited from procedure catalog.
        specialty:
          $ref: '#/components/schemas/Specialty'
          description: Specialty classification inherited from procedure catalog.
        laterality_support:
          type: boolean
          description: Whether bilateral billing is applicable for this code.
        quantity_selected:
          type: integer
          minimum: 1
          default: 1
          description: Number of segments/vertebrae/structures treated (only if billing_mode allows).
        laterality:
          $ref: '#/components/schemas/Laterality'
          default: UNILATERAL
          description: Whether this procedure is billed unilaterally or bilaterally.
    BillingModifiers:
      type: object
      description: >
        Spine-specific billing variables. Only applicable fields based on selected procedures' billing modes are used.
        Omit fields that do not apply; defaults (quantity_selected=1, laterality=UNILATERAL) are used.
      properties:
        quantity_selected:
          type: integer
          minimum: 1
          default: 1
          description: Number of segments/vertebrae/structures treated (only if billing_mode allows).
        laterality:
          $ref: '#/components/schemas/Laterality'
          default: UNILATERAL
          description: Whether procedure is billed unilaterally or bilaterally (only if supported).
        vertebral_region:
          type: string
          enum: [cervical, thoracic, lumbar, lumbosacral, sacral]
          description: Anatomical region (informational metadata, no billing effect).
        surgical_approach:
          type: string
          enum: [anterior, posterior, posterolateral, lateral, minimally_invasive]
          description: Surgical technique (informational metadata, no billing effect).
        fusion_status:
          type: string
          enum: [decompression_only, fusion, hybrid]
          description: Fusion configuration (informational metadata, no billing effect).
        implant_category:
          type: string
          enum: [graft, hardware, arthroplasty, neural_device]
          description: Type of implant (informational metadata, no billing effect).
        osteoporosis_aware:
          type: boolean
          description: Whether osteoporosis is documented (informational metadata, no billing effect).
        clinical_context:
          type: string
          enum: [traumatic, degenerative, neoplastic, infectious, deformity, revision]
          description: Clinical diagnosis context (informational metadata, no billing effect).
    CalculateRequest:
      type: object
      required: [selected_codes, auxiliaries_count, requires_anesthesia, access_route_type]
      properties:
        selected_codes:
          type: array
          minItems: 1
          items:
            $ref: '#/components/schemas/SelectedCode'
          description: Physician-selected CBHPM codes.
        auxiliaries_count:
          type: integer
          minimum: 0
          maximum: 4
          description: Number of auxiliary surgeons (0–4).
        requires_anesthesia:
          type: boolean
          description: Whether an anesthesiologist is required.
        access_route_type:
          $ref: '#/components/schemas/AccessRouteType'
        urgency_emergency:
          type: boolean
          description: >
            When true, applies the CBHPM item 2 urgency/emergency 30% surcharge
            to all medical fees (surgeon, auxiliaries, anesthesiologist).
            Applicable for acts performed 19h–7h, on weekends, or on holidays.
        modifiers:
          $ref: '#/components/schemas/BillingModifiers'
          description: Spine-specific billing variables (optional).
        adjustments:
          type: array
          items:
            type: string
          description: >
            Array of CBHPM adjustment codes to apply to the calculation.
            Valid codes: emergency_special_hours, pediatric_low_weight_or_premature,
            pediatric_neonate_or_infant, pediatric_child_under_12.
            Percentages are additive (e.g., two 30% adjustments = 60%, not 69%).
    CodeBreakdown:
      type: object
      required: [cbhpm_code, description, porte, base_value, is_principal, billing_mode, quantity_selected, quantity_multiplier, laterality, laterality_multiplier, adjusted_value]
      properties:
        cbhpm_code:
          type: string
        description:
          type: string
        porte:
          type: string
        base_value:
          type: number
          format: float
          description: Full monetary value of this code's porte (before any discount).
        is_principal:
          type: boolean
          description: True for the highest-value code selected as the principal procedure.
        billing_mode:
          $ref: '#/components/schemas/BillingMode'
          description: Billing mode that determines quantity multiplier.
        quantity_selected:
          type: integer
          description: Number of segments/vertebrae/structures selected (1 if not applicable).
        quantity_multiplier:
          type: number
          format: float
          description: Multiplier applied based on quantity (e.g., 1.0, 2.0, 3.0).
        laterality:
          $ref: '#/components/schemas/Laterality'
          description: Whether billed unilaterally or bilaterally.
        laterality_multiplier:
          type: number
          format: float
          description: Multiplier applied based on laterality (1.0 or 2.0).
        adjusted_value:
          type: number
          format: float
          description: base_value × quantity_multiplier × laterality_multiplier (before CBHPM discounts).
    AuxiliaryFee:
      type: object
      required: [position, percentage, fee]
      properties:
        position:
          type: integer
          description: Auxiliary position (1 = first, 2 = second, etc.).
        percentage:
          type: number
          format: float
          description: Percentage applied (60, 40, 30, or 30).
        fee:
          type: number
          format: float
          description: Calculated fee for this auxiliary.
    AppliedAdjustment:
      type: object
      required: [code, label, percentage, source]
      properties:
        code:
          type: string
          description: Adjustment code identifier.
        label:
          type: string
          description: Human-readable label for the adjustment.
        percentage:
          type: number
          format: float
          description: Percentage value of this adjustment.
        source:
          type: string
          description: Source reference (e.g., CBHPM manual section).
    SurgeonBreakdown:
      type: object
      required: [principal_value, additional_gross, discount_rate, additional_discounted, surgeon_total]
      properties:
        principal_value:
          type: number
          format: float
          description: Monetary value of the principal (highest-value) procedure.
        additional_gross:
          type: number
          format: float
          description: Sum of all non-principal procedure values before applying the CBHPM discount.
        discount_rate:
          type: number
          format: float
          description: Discount rate applied to additional procedures (0.50 for same route, 0.70 for different routes).
        additional_discounted:
          type: number
          format: float
          description: Sum of additional procedure values after applying the discount.
        surgeon_total:
          type: number
          format: float
          description: Final surgeon fee (principal_value + additional_discounted).
    CalculateResponse:
      type: object
      required:
        - code_breakdown
        - access_route_type
        - surgeon_breakdown
        - lead_surgeon_fee
        - individual_auxiliary_fees
        - auxiliaries_fee
        - anesthesiologist_fee
        - final_total
        - total_base
        - selected_adjustments
        - total_adjustment_percentage
        - adjustment_value
      properties:
        code_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/CodeBreakdown'
          description: Per-code contribution with principal flag.
        access_route_type:
          $ref: '#/components/schemas/AccessRouteType'
        surgeon_breakdown:
          $ref: '#/components/schemas/SurgeonBreakdown'
        lead_surgeon_fee:
          type: number
          format: float
          description: Final surgeon fee (equals surgeon_breakdown.surgeon_total).
        individual_auxiliary_fees:
          type: array
          items:
            $ref: '#/components/schemas/AuxiliaryFee'
          description: Per-auxiliary fee details (CBHPM 5.1 – 60/40/30/30%).
        auxiliaries_fee:
          type: number
          format: float
          description: Sum of all individual auxiliary fees.
        anesthesiologist_fee:
          type: number
          format: float
          description: Fixed anesthesiologist fee (if applicable).
        final_total:
          type: number
          format: float
          description: Grand total (lead_surgeon_fee + auxiliaries_fee + anesthesiologist_fee).
        total_base:
          type: number
          format: float
          description: Simple sum of all selected code porte values (before discounting). Retained for compatibility.
        urgency_emergency_applied:
          type: boolean
          description: Whether the CBHPM item 2 urgency/emergency surcharge was applied.
        urgency_emergency_percentage:
          type: number
          format: float
          description: Surcharge percentage (30.0 when applied, 0 otherwise).
        urgency_emergency_value:
          type: number
          format: float
          description: Absolute monetary value of the surcharge added to the total.
        selected_adjustments:
          type: array
          items:
            $ref: '#/components/schemas/AppliedAdjustment'
          description: Array of CBHPM adjustments applied to this calculation.
        total_adjustment_percentage:
          type: number
          format: float
          description: Combined percentage from all applied adjustments (additive, not multiplicative).
        adjustment_value:
          type: number
          format: float
          description: Absolute monetary value of all combined adjustments.
...